Move Uniforms and Shader util classes to regl-utils.js
